Cache curatedGenes/CancerGenes/Info
