+++ core/src/main/java/org/mskcc/cbio/oncokb/util/CacheUtils.java
@@ -54,6 +54,8 @@ public class CacheUtils {
 
     private static Map<String, Long> recordTime = new HashedMap();
 
+    private static Info oncokbInfo;
+
     private static Observer numbersObserver = new Observer() {
         @Override
         public void update(Observable o, Object arg) {
@@ -211,6 +213,7 @@ public class CacheUtils {
             cacheDownloadAvailability();
             System.out.println("Cached downloadable files availability on github: " + MainUtils.getTimestampDiff(current) + " at " + MainUtils.getCurrentTime());
 
+            oncokbInfo = ApplicationContextSingleton.getInfoBo().get();
             System.out.println("Cached oncokb info " + MainUtils.getTimestampDiff(current) + " at " + MainUtils.getCurrentTime());
 
             registerOtherServices();
@@ -238,7 +241,7 @@ public class CacheUtils {
     }
 
     public static Info getInfo() {
-        return ApplicationContextSingleton.getInfoBo().get("test");
+        return oncokbInfo;
     }
 
     public static Boolean containGeneByEntrezId(Integer entrezId) {
