+++ web/src/main/java/org/mskcc/cbio/oncokb/api/pub/v1/InfoApiController.java
@@ -1,7 +1,8 @@
 package org.mskcc.cbio.oncokb.api.pub.v1;
 
+import org.mskcc.cbio.oncokb.cache.CacheFetcher;
 import org.mskcc.cbio.oncokb.model.OncoKBInfo;
-import org.mskcc.cbio.oncokb.util.*;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -12,8 +13,11 @@ import org.springframework.stereotype.Controller;
  */
 @Controller
 public class InfoApiController implements InfoApi {
+    @Autowired
+    CacheFetcher cacheFetcher;
+
     @Override
     public ResponseEntity<OncoKBInfo> infoGet() {
-        return new ResponseEntity<>(new OncoKBInfo(), HttpStatus.OK);
+        return new ResponseEntity<>(this.cacheFetcher.getOncoKBInfo(), HttpStatus.OK);
     }
 }
