--- web/src/main/java/org/mskcc/cbio/oncokb/config/cache/CacheKeys.java
@@ -1,5 +1,0 @@
-package org.mskcc.cbio.oncokb.config.cache;
-
-public final class CacheKeys {
-    public static final String CANCER_GENE_LIST = "cancerGeneList";
-}
